Update README.md on installation instructions
Dropped the advice to just download `index.js` and `jump.kak` files, because this will also require renaming it to `.mjs` extension, and then we'd need to introduce more error checking and it's so tedious it's not really worth it. I guess people who really want to avoid cloning the whole repo can figure out these Node intricacies on their own.